+++ packages/ui/src/dialog/close-icon.tsx
@@ -22,6 +22,7 @@ const CloseIcon = forwardRef< HTMLButtonElement, CloseIconProps >(
 						{ ...props }
 						icon={ icon ?? close }
 						label={ label ?? __( 'Close' ) }
+						data-wp-dialog-close-icon=""
 					/>
 				}
 			/>